Create remove-me.py
Take your name off of FigShare. This is for when you use a program and API to upload things and accidentally get an Author Credit for the items you uploaded. 

This is because I accidentally got like 1,200 author credits and my institution wasn't happy about seeing my name attached to things my software updated.